+++ FlexEngine/src/Managers/CartManager.cpp
@@ -199,7 +199,7 @@ namespace flex
 								if (newCartChainID != InvalidCartChainID &&
 									newCartChainID >= m_CartChains.size())
 								{
-									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", newCartChainID, m_CartChains.size());
+									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", newCartChainID, (u32)m_CartChains.size());
 								}
 							}
 							else if (c1 == InvalidCartChainID)
@@ -219,7 +219,7 @@ namespace flex
 								if (m_Carts[i]->chainID != InvalidCartChainID &&
 									m_Carts[i]->chainID >= m_CartChains.size())
 								{
-									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", m_Carts[i]->chainID, m_CartChains.size());
+									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", m_Carts[i]->chainID, (u32)m_CartChains.size());
 								}
 							}
 							else if (c2 == InvalidCartChainID)
@@ -239,7 +239,7 @@ namespace flex
 								if (m_Carts[j]->chainID != InvalidCartChainID &&
 									m_Carts[j]->chainID >= m_CartChains.size())
 								{
-									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", m_Carts[j]->chainID, m_CartChains.size());
+									PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", m_Carts[j]->chainID, (u32)m_CartChains.size());
 								}
 							}
 							else
@@ -328,7 +328,7 @@ namespace flex
 				if (cart->chainID != InvalidCartChainID &&
 					cart->chainID >= m_CartChains.size())
 				{
-					PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", cart->chainID, m_CartChains.size());
+					PrintError("Cart chain update failed! Cart has invalid cart chain ID: %u, num cart chains: %u\n", cart->chainID, (u32)m_CartChains.size());
 				}
 			}
 		}
